+++ crosswrd-app/src/EditLights.tsx
